Create Read me
Step 1: Download and open the required files
Download and open the Excel workbook Sample.xlsx. This workbook consists of a sheet called Sheet1 that contains financial data. It also contains a sheet called Contacts, which you will use in another part of this exercise. A preview of Sheet1 is visible in the screenshot. You will complete this exercise by adding data to and formatting, Sheet1 in the workbook.

Workbook with Sheet1 and Contacts sheets. Content in Sheet1 displayed.
Next, download the Exchange Rates.pdf  file. You will add the data from this PDF to your workbook. Save the PDF file in a location where you can quickly find it to use in Step 4.

Step 2: Verify and correct the existing information in the sheet
To begin, open Sheet1 in the Sample Excel file, then carry out the following tasks:

The State Abbreviation column is unnecessary and can be deleted because the sheet contains international data. Select column B, then right-click and select Delete Columns from the shortcut menu.

The entries in column D appear different than the other columns. This is because the column is too narrow. Position the mouse pointer on the vertical line between the D and E column identifiers, then double-click to resize the column automatically.

The figures in the spreadsheet will be a key focus of the meeting, so you need to confirm that there are no other issues in the number range. The number in cell G18 is sitting over to the left of the cell, which indicates that Excel is treating the entry as text. The ~ character at the beginning of the entry is causing the problem, causing the entry to be auto-detected as text. Double-click on cell G18 to enter Edit mode. Then move the cursor to the right of the character and press backspace. Press Enter to confirm the amendment and correctly realign the entry.

Add a column to hold country data to display the worksheet’s international financial data. Add this new column to the right of the State column and fill in the country entries. Select the letter C that identifies the column to select it. Then right-click and choose Insert Columns from the shortcut menu. A new blank column appears to the left of the column titled Office Type.

Type the heading Country in C1 and then USA in C2. Move the cursor back to C2, then hover the mouse pointer over the bottom right-hand corner until it becomes the narrow black cross. Hold down the mouse button and drag it down as far as row 21 to copy the entry quickly into the appropriate cells. Add Japan to C22 and use a similar Autofill technique to complete the country entries in rows 23 and 24. Add Germany to C25 and use a similar Autofill technique to complete the country entries in rows 26 and 27.

Renee would like the information relating to Japan and Germany to be easily identifiable. So, she has asked you to add these country names as headings above both sets of information to create a clearer distinction between the data for these countries and the USA. First, select row 22 on the left-hand side. Then right-click and choose Insert Rows from the shortcut menu. Select row 26 and repeat the insert rows operation. Then type Japan in cell A22 and Germany in A26.

Step 3: Format the data
The worksheet now contains all the required information. Your next step is to improve its readability before presenting it to management.

First, you need to emphasize the titles in row 1. Select the range of cells A1 to H1, and on the Home tab, select the Font group. Then choose font size 14 and a different background color. Without canceling the selection, center the headings. On the Home tab, select the Alignment group, then select the Center choice. 

The heading Revenue in E1 and the Depreciation heading in H are only partially visible, as the columns are not wide enough to accommodate the text. Double-click on the vertical separator between E and F to automatically resize column E. Then perform a similar adjustment on column H.

The titles in F1 and G1 are also exhibiting visibility issues. However, widening the columns would create too much visible white space. Instead, select F1 and G1. On the Home tab, select the Alignment group, then choose Wrap Text. 

Column G still needs to be a little wider. So, hover the mouse pointer over the dividing line between the column letters G and H, hold down the mouse button, and drag it slightly to the right.

The headings in the first row now stand out and are easier to read. You now need to emphasize the country headings in A22 and A26. Move the cursor to A1. On the Home tab, in the Clipboard group, select the Format Painter button and then select A22. Repeat the process for A26. Next, select cells B22 to H22 and apply the same background color. Repeat this process for cells B26 to H26.

Next, you need to add currency formats to distinguish between the three currencies in the sheet. Select cells E2 to H21. On the Home tab in the Number group, select the currency drop-down and then choose the dollar symbol. However, this action generates extra characters and decimal places. So, column F needs to be resized a little.

Select cells E27 to H29 and choose the Euro currency format from the Currency drop-down.

Finally, select cells E23 to H25 and select the Currency drop-down again. There is no Japanese Yen choice offered in the immediate list. Select the More Accounting Formats choice at the bottom. Select the Currency category on the left of the number format dialog. Choose ¥ Japanese (also known as Japanese Yen) from the Symbol drop-down on the right.

Step 4: Add new information to the workbook
Now that you’ve improved the formatting and readability of the data in Sheet 1, you need to add the information from the Exchange Rates.pdf  file to the Microsoft Excel workbook.

The information on exchange rates needs to be in the Excel file but held separately from the data you have just been working on. Create a new worksheet by selecting the + symbol in the worksheet tab area. 

In cell A1 of the new sheet, type the heading USD Amount. In cell B1, type the heading Currency. In cell C1 type the heading Rate. In A2, type the number 1, and in B2 type Yen. In A3, type the number 1, and in B3 type Euro. Then, in C2 and C3, add the appropriate exchange rate figures from the PDF.

To provide visual consistency between both sheets, format these headings to match the headings in Sheet1.

Step 5: Manage the worksheets
The workbook is now almost complete. In the last few steps, you must name and organize the worksheets you need for the presentation.

The original worksheet that you edited is still called Sheet1. The sheet you added to hold the exchange rate information will also have a generic title of Sheet followed by a number. Both titles could be more descriptive. So, rename the first sheet Sample Figures, and the other sheet Exchange Rates. Right-click the relevant sheet tab, type the new names, and select OK. 

To prepare for your presentation, it is best to position the Exchange Rates sheet to the right of the Sample Figures sheet. Point your cursor at the Exchange Rates tab, hold down the mouse button, and drag it to its new position.

You don’t yet have all the required data for the Contacts sheet. In addition, this sheet is not directly relevant to your presentation. However, the information will be needed in the future. Rather than deleting it, you can hide it for the duration of the presentation. Select the sheet. On the Home tab in the Cells group, choose Format and then select Hide & Unhide to hide the sheet.

Conclusion
You have now prepared the worksheet for your presentation by adding the required data and improving its readability. By completing this exercise, you have gained experience in editing a Microsoft Excel file and formatting it to a professional standard. As you worked through the steps, you used techniques that will be useful to you as you work with Excel.